refactor: Return AddressId from Create method
Refactor to try to return the AddressId with the Create method. Refactor the Main program to run CombinedExample, which chains Create - Get - Update - Delete to show that all CRUD functionality works.

This is currently untested, in particular the SQL query in Create with an output parameter. Some resources I saw online suggest that this may not work unless the command is registered as a stored procedure.